requirements.txt
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.8.17
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.2
django-filer==1.2.6
django-formtools==2.0
django-mptt==0.8.7
django-polymorphic==1.0.2
django-sekizai==0.10.0
Django-Select2==4.3.2
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.1.2
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.0.2
djangocms-inherit==0.2.2
djangocms-installer==0.9.3
djangocms-link==2.1.0
djangocms-picture==2.0.4
djangocms-snippet==1.9.2
djangocms-style==2.0.1
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
easy-thumbnails==2.3
html5lib==0.9999999
olefile==0.44
Pillow==4.0.0
pytz==2016.10
six==1.10.0
tzlocal==1.3
Unidecode==0.4.20
virtualenv==15.1.0
gunicorn==19.6.0
psycopg2
whitenoise
